Refactor temperature handling in CurveManager and Line structures; improve safety in GraphRenderer calculations

+++ include/CurveManager.h
@@ -117,8 +117,8 @@ public:
     {
         return isSkipDown(currentSegmentIndex);
     }
-    unsigned long getSegmentStartTemperature() { return getSegmentStartTemperature(currentSegmentIndex); }
-    unsigned long getSegmentStartTemperature(int index) { return index == 0 ? 20 : originalCurve.elems[index - 1].endTemp; }
+    float getSegmentStartTemperature() { return getSegmentStartTemperature(currentSegmentIndex); }
+    float getSegmentStartTemperature(int index) { return index == 0 ? 20.0f : originalCurve.elems[index - 1].endTemp; }
     float getHeatingSpeed() const;
     float getDeltaTemp() const
     {
